Migrated from Bitbucket to GitHub
We have migrated the source code repository for ChiToolbox from Bitbucket to GitHub.
One of the reasons for this is the potential for a DOI from Zenodo using GitHub. Bitbucket have indicated this is not on their roadmap. There is also greater visibility for projects on GitHub, where Bitbucket seems to have a more corporate focus.
A number of changes will need to be made. This commit changes the URLs and mentions of Bitbucket to the GitHub equivalents. Other changes will take a little more configuration.
The wiki pages are still on Bitbucket, so those links remain for now. These will be migrated to GitHub pages, and the source for that will be in a /docs folder.